Split setting tests into multiple files
